Create `grades' module for Grade classes
To prevent want seems to be an import cycle begging to happen, move the
definitions of PartGrade and AssignmentComponentGrade to their own
module, grades.
